Add files
Scripts from my final project. The game was a simple two-player space game called "Space Tanks... in Space". The objective is simple. Survive for as long as possible as enemies spawn at increasing speed. Enemies can only be destroyed by bullets of the same color, requiring the player(s) to change color to defeat them. The unique mechanic of the game is that enemies will spawn in six different colors (red, blue, green, yellow, cyan, and magenta) while players can only change to between three (red, blue, and green). For the remaining three colors, the players must coordinate and collide to gain access to the combination colors. The game was very well received during the peer review demonstration and was given an A.